requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==2.1.0
    # via mediapipe
attrs==24.2.0
    # via mediapipe
cachetools==5.5.0
    # via tox
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via sounddevice
chardet==5.2.0
    # via tox
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tox
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distlib==0.3.8
    # via virtualenv
filelock==3.16.1
    # via
    #   tox
    #   virtualenv
flatbuffers==24.3.25
    # via mediapipe
fonttools==4.54.1
    # via matplotlib
idna==3.10
    # via requests
jax==0.4.33
    # via mediapipe
jaxlib==0.4.33
    # via
    #   jax
    #   mediapipe
kiwisolver==1.4.7
    # via matplotlib
matplotlib==3.9.2
    # via mediapipe
mediapipe==0.10.15
    # via -r requirements.in
ml-dtypes==0.5.0
    # via
    #   jax
    #   jaxlib
numpy==1.26.4
    # via
    #   contourpy
    #   jax
    #   jaxlib
    #   matplotlib
    #   mediapipe
    #   ml-dtypes
    #   opencv-contrib-python
    #   scipy
opencv-contrib-python==4.10.0.84
    # via
    #   -r requirements.in
    #   mediapipe
opt-einsum==3.4.0
    # via jax
packaging==24.1
    # via
    #   matplotlib
    #   pyproject-api
    #   tox
pillow==10.4.0
    # via matplotlib
platformdirs==4.3.6
    # via
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via tox
protobuf==4.25.5
    # via mediapipe
pycparser==2.22
    # via cffi
pyparsing==3.1.4
    # via matplotlib
pyproject-api==1.8.0
    # via tox
python-dateutil==2.9.0.post0
    # via matplotlib
requests==2.32.3
    # via -r requirements.in
scipy==1.14.1
    # via
    #   jax
    #   jaxlib
six==1.16.0
    # via python-dateutil
sounddevice==0.5.0
    # via mediapipe
tox==4.21.1
    # via -r requirements.in
urllib3==2.2.3
    # via requests
virtualenv==20.26.6
    # via tox
